backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.6.0
python-socketio==5.16.0
pymongo==4.5.0
python-dotenv==1.0.0
dnspython==2.4.2
